OF: base: fix infinite loop in of_find_node_with_property
of_find_node_with_property suffers from infinite looping, because it
does not check for root node of the tree iterated over. This adds a
check for parent pointer of last tested iterator entry, which is
NULL for the root node. Also set from node to current root_node if
NULL.

Signed-off-by: Sebastian Hesselbarth <sebastian.hesselbarth@gmail.com>